Move some methods from the helpers file into the versioning file
